refactor: update import paths for shape components and clean up unused code
