+++ app/components/Shapes/FigureBar/figurebar-item.tsx
