+++ app/components/Shapes/shape/types/diamond.tsx
