+++ app/components/Shapes/shape/types/hexagon.tsx
