+++ app/components/Shapes/shape/types/triangle.tsx
